fix(ESSNTL-3781): Fixes filter issues in options object
